fixed bugs & contacts ssdb

+++ doc/db/postgresql/function/delete_contact.sql
@@ -1,23 +1,34 @@
 CREATE OR REPLACE FUNCTION delete_contact(a_id INTEGER,
                                           b_id INTEGER)
-RETURNS VOID AS
+RETURNS INTEGER AS
 $$
 DECLARE
-    now contacts.created_at%TYPE;
+    count INTEGER;
+    now   contacts.created_at%TYPE;
 BEGIN
-    now = now();
+    SELECT count(c.user_id)
+    INTO count
+    FROM contacts c
+    WHERE c.user_id = delete_contact.a_id
+    AND c.contact_id = delete_contact.b_id;
 
-    DELETE FROM contacts c
-    WHERE   (c.user_id = a_id AND c.contact_id = b_id)
-    OR      (c.user_id = b_id AND c.contact_id = a_id);
+    IF count <> 0 THEN
+        now = now();
 
-    UPDATE  users u
-    SET     contact_version = u.contact_version + 1,
-            updated_at = now
-    WHERE   u.id = a_id
-    OR      u.id = b_id;
+        DELETE FROM contacts c
+        WHERE   (c.user_id = a_id AND c.contact_id = b_id)
+        OR      (c.user_id = b_id AND c.contact_id = a_id);
 
-    RETURN;
+        UPDATE  users u
+        SET     contact_version = u.contact_version + 1,
+                updated_at = now
+        WHERE   u.id = a_id
+        OR      u.id = b_id;
+
+        RETURN 0;
+    ELSE
+        RETURN 1;
+    END IF;
 END;
 $$
 LANGUAGE plpgsql;
